Pass Jolt interfaces through the API as well.
