+++ api/Cargo.toml
@@ -20,6 +20,7 @@ anyhow = "1.0"
 zstd = { version = "0.12", default-features = false }
 
 nexus-vm = { path = "../vm" }
+nexus-jolt = { path = "../jolt" }
 nexus-nova = { path = "../nova", features = ["spartan"] }
 nexus-tui = { path = "../tools/tui" }
 spartan = { path = "../spartan", package = "ark-spartan" }
