+++ api/examples/prover_run.rs
@@ -41,10 +41,12 @@ fn main() {
 
     println!("Proving execution of length {}...", trace.blocks.len());
     let tr = prover::nova::init_circuit_trace(trace).expect("error constructing circuit");
-    
-    let mut proof = prover::nova::prove_seq_step(None, &public_params, &tr).expect("error proving execution");
+
+    let mut proof =
+        prover::nova::prove_seq_step(None, &public_params, &tr).expect("error proving execution");
     for _ in 1..tr.steps() {
-        proof = prover::nova::prove_seq_step(Some(proof), &public_params, &tr).expect("error proving execution");
+        proof = prover::nova::prove_seq_step(Some(proof), &public_params, &tr)
+            .expect("error proving execution");
     }
 
     print!("Verifying execution...");
