+++ api/src/prover/mod.rs
@@ -1,2 +1,2 @@
+pub mod jolt;
 pub mod nova;
-//pub mod jolt;
